requirements.txt
instructor==1.3.3
munch==4.0.0
openai==1.34.0
Requests==2.32.3
weather_ch==0.0.4
